refactor: boundary drift 정리
