+++ src/entities/DeadCode/lib/buildDeadCodeTree.ts
