+++ src/entities/DeadCode/model/atoms.ts
