+++ src/features/Canvas/lib/pruneDetachedNodes.ts
