+++ src/features/Code/CodeViewer/ui/CodeToken.tsx
@@ -1,11 +1,11 @@
 import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import type React from 'react';
 import { fullNodeMapAtom } from '@/entities/AppView/model/atoms';
+import { pruneDetachedNodes } from '@/features/Canvas/lib/pruneDetachedNodes';
 import { cardPositionsAtom, transformAtom, visibleNodeIdsAtom } from '@/features/Canvas/model/atoms';
 import { getTokenStyle } from '@/features/Code/CodeViewer/core/styler/tokenStyles';
 import { useGotoDefinition } from '@/features/Code/CodeViewer/lib/useGotoDefinition';
 import { activeLocalVariablesAtom } from '@/features/Code/FocusMode/model/atoms';
-import { pruneDetachedNodes } from '@/pages/shared/features/Canvas/lib/pruneDetachedNodes';
 
 const CodeToken = ({
   text,
