+++ src/features/File/OpenFiles/ui/IDEScrollView.tsx
@@ -10,8 +10,8 @@ import { useAtomValue } from 'jotai';
 import { useEffect, useMemo } from 'react';
 import { filesAtom, fullNodeMapAtom } from '@/entities/AppView/model/atoms';
 import { getSelectedDeadCodeFilePaths, getSelectedDeadCodeLinesByFile } from '@/entities/DeadCode/lib/computed';
+import { deadCodeResultsAtom, selectedDeadCodeItemsAtom } from '@/entities/DeadCode/model/atoms';
 import { activeTabAtom, openedTabsAtom } from '@/features/File/OpenFiles/model/atoms.ts';
-import { deadCodeResultsAtom, selectedDeadCodeItemsAtom } from '@/pages/shared/features/DeadCode/model/atoms.ts';
 import { useScrollNavigation } from '../lib/useScrollNavigation.ts';
 import FileSection from './FileSection.tsx';
