+++ src/pages/canvas/features/CodeCard/CodeCard.tsx
@@ -2,6 +2,7 @@ import { useAtomValue, useSetAtom } from 'jotai';
 import { useEffect, useMemo, useRef } from 'react';
 import { filesAtom } from '@/entities/AppView/model/atoms';
 import type { CanvasNode } from '@/entities/CanvasNode/model/types';
+import { deadCodeResultsAtom } from '@/entities/DeadCode/model/atoms';
 import { getFoldableLinesByMaxDepth } from '@/features/Code/CodeFold/lib/foldUtils';
 // Atoms
 import { foldedLinesAtom } from '@/features/Code/CodeFold/model/atoms';
@@ -12,7 +13,6 @@ import { renderVueFile } from '@/features/Code/CodeViewer/core/renderer/renderVu
 import { hoveredFilePathAtom } from '@/features/Code/CodeViewer/model/atoms';
 import VueTemplateSection from '@/features/Code/CodeViewer/ui/VueTemplateSection';
 import { activeTabAtom } from '@/features/File/OpenFiles/model/atoms';
-import { deadCodeResultsAtom } from '@/pages/shared/features/DeadCode/model/atoms';
 import CodeCardCopyButton from './ui/CodeCardCopyButton';
 // UI Components
 import CodeCardHeader from './ui/CodeCardHeader';
