+++ src/pages/dead-code/features/DeadCodeSidebar/lib/useDeadCodeCategories.ts
@@ -1,8 +1,8 @@
 import { useAtomValue } from 'jotai';
 import { useMemo } from 'react';
 import { getDeadCodeCategories } from '@/entities/DeadCode/lib/computed';
+import { deadCodeResultsAtom } from '@/entities/DeadCode/model/atoms';
 import type { DeadCodeCategoryInfo } from '@/entities/DeadCode/model/types';
-import { deadCodeResultsAtom } from '@/pages/shared/features/DeadCode/model/atoms.ts';
 
 export function useDeadCodeCategories(): DeadCodeCategoryInfo[] {
   const deadCodeResults = useAtomValue(deadCodeResultsAtom);
