+++ src/pages/dead-code/features/DeadCodeSidebar/ui/renderCategoryIcon.tsx
